requirements.txt
# Polymarket Copy Trading Bot - Dependencies
# ==========================================

# Polymarket Official SDK
py-clob-client>=0.17.0

# Web3 & Blockchain
web3>=6.0.0
eth-account>=0.8.0

# HTTP & WebSocket
requests>=2.31.0
websockets>=12.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# GUI Dashboard
customtkinter>=5.2.0
Pillow>=10.0.0
tkcalendar>=1.6.1

# Charts & Visualization  
matplotlib>=3.7.0
plotly>=5.18.0

# Database
pymongo>=4.6.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async Support
asyncio-throttle>=1.0.2

# Logging & Utils
colorlog>=6.8.0
tabulate>=0.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
